исправления

+++ src/main/java/ru/plenkkovii/weather/dto/LocationViewResponseDTO.java
@@ -8,8 +8,8 @@ import lombok.Data;
 public class LocationViewResponseDTO {
     private String name;
     private String country;
-    private double temp;
-    private double feelsLike;
+    private int temp;
+    private int feelsLike;
     private String weather;
     private int humidity;
 }
